refactor(text-field): move out re-used components to Field
